added bounds to limit cycle continuation in pycont
the bounds only seemed to work for pycont when not doing limit cycle continuation. I can't find any reference to StopAtPoints or TestFuncs in the LimitCycleCurve continuation class.
I have added rl0 and rl1 (lower and upper bounds on parameter) into the data in auto.c and added code to set it up using automod.c and SetData